[analyzer] Add the initial state for the analysis
This diff introduces the basic state setup for our analysis.

The basic analysis will only work in *input parameters and parameters
returned by `_create` functions*. For the basic implementation we will
not look into structures.

This diff does not do any analysis yet.

Test:
`llvm-lit -sv clang/test/Analysis/object-count-checks.c` passes.

* RefCount

This is our state tracking for our static analyzer. It will increment
with acquire calls and decrement with release calls. Each object
will map directly to one of these RefCount states.

It is fairly empty now but it will extend to include autorelease count
and whether or not it has escaped the execution context.

* NoAnalysisAnnotation

Introduce an annotation to bootstrap the object count as well as be an
escape hatch for cases we did not handle.

* checkBeginFunction

This is the main introduced function. It will run for every top level
function analysis. It only operations on regular functions (no
constructor, destructors, etc). It will initialize the object state
for single pointer types (e.g. foo_t*, header_t*). This is because we
currently have no way of identifying our objects.

If the pointer type has the consumed attribute we initialize it as an
owned object, if it doesn't we initialize it as unowned.