[analyzer] Add Object ReleaseNotOwned Checker
This diff introduces the first report type ReleaseNotOwned. This happens
when you release a reference you don't own (i.e. protects against
double-release).

We leverage the report type in RetainCountDiagnostics to share the code.
We don't completely merge our analysis just to stay independent.

Current limitation is we don't have acquire or `_create`.

As a feature, we NEVER touch global variables or variables within
structs. We can't reason through whether it's valid or not and thus
don't do any analysis on it.